requirements.txt
minio==7.1.15
pymongo==4.5.0
pandas==2.0.3
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0